pyspotter/requirements.txt
timezonefinder==5.2.0
requests==2.24.0
pytz==2020.1
beautifulsoup4==4.9.3
python_dateutil==2.8.2
